Merged PR 251: ZeroBounce API Integration
Implements the ZeroBounce API integration for all leads received through our Azure API service. This process runs AFTER the response has been returned to the lead vendor, which means that we may flag a lead as non-billable after we have told the vendor the lead is billable. This process first checks to see if we have performed a lookup on the lead email with ZeroBounce in the last 90 days. If we have, we use the result of that API request to evaluate the business rule, which is only "valid" email addresses will be accepted and will trigger a welcome email. When we make a request to the ZeroBounce API, we store the result in dbo.ZeroBounce regardless of the outcome. If the lead fails the validation, we also add a dbo.PartnerContactLeadStatus to indicate the failure (Name = "Verification Failure").

This pull request also addresses a bug with our tracking api endpoint which was not supporting the new "tinyId" parameter for campaign tracking.

Related work items: #716, #719