Merged PR 292: Less aggressive deletion logic for job data mining process
I suspect that there is a problem with Aerotek and TEK's job search api (or the way that we are interacting with it). When crawling these job sites, there are always a bunch of job pages that don't show up that are still active pages. To address this discrepancy, we will do the following:

Rather than eliminating jobs on the basis that they don't exist in the job site's search results, make a request to each job page that we suspect may have been deleted. If the page returns a valid response that does not contain a "job missing" message, leave it alone (don't delete the job). Otherwise, delete the job.

Related work items: #802